+++ lib/bootevent.js
@@ -14,9 +14,13 @@ function boot (env) {
     next();
   }
 
+  function hasBootErrors(ctx) {
+    return ctx.bootErrors && ctx.bootErrors.length > 0;
+  }
+
   function setupMongo (ctx, next) {
 
-    if (ctx.bootErrors && ctx.bootErrors.length > 0) {
+    if (hasBootErrors(ctx)) {
       return next();
     }
 
@@ -36,12 +40,25 @@ function boot (env) {
     }
   }
 
-  function setupInternals (ctx, next) {
-    if (ctx.bootErrors && ctx.bootErrors.length > 0) {
+  function setupAuthorization (ctx, next) {
+    if (hasBootErrors(ctx)) {
       return next();
     }
 
     ctx.authorization = require('./authorization')(env, ctx);
+    ctx.authorization.storage.reload(function loaded (err) {
+      if (err) {
+        ctx.bootErrors = ctx.bootErrors || [ ];
+        ctx.bootErrors.push({'desc': 'Unable to setup authorization', err: err});
+      }
+      next();
+    });
+  }
+
+  function setupInternals (ctx, next) {
+    if (hasBootErrors(ctx)) {
+      return next();
+    }
 
     ///////////////////////////////////////////////////
     // api and json object variables
@@ -68,7 +85,7 @@ function boot (env) {
   }
 
   function ensureIndexes (ctx, next) {
-    if (ctx.bootErrors && ctx.bootErrors.length > 0) {
+    if (hasBootErrors(ctx)) {
       return next();
     }
 
@@ -83,7 +100,7 @@ function boot (env) {
   }
 
   function setupListeners (ctx, next) {
-    if (ctx.bootErrors && ctx.bootErrors.length > 0) {
+    if (hasBootErrors(ctx)) {
       return next();
     }
 
@@ -118,7 +135,7 @@ function boot (env) {
   }
 
   function setupBridge (ctx, next) {
-    if (ctx.bootErrors && ctx.bootErrors.length > 0) {
+    if (hasBootErrors(ctx)) {
       return next();
     }
 
@@ -130,7 +147,7 @@ function boot (env) {
   }
 
   function setupMMConnect (ctx, next) {
-    if (ctx.bootErrors && ctx.bootErrors.length > 0) {
+    if (hasBootErrors(ctx)) {
       return next();
     }
 
@@ -142,7 +159,7 @@ function boot (env) {
   }
 
   function finishBoot (ctx, next) {
-    if (ctx.bootErrors && ctx.bootErrors.length > 0) {
+    if (hasBootErrors(ctx)) {
       return next();
     }
 
@@ -154,6 +171,7 @@ function boot (env) {
   return require('bootevent')( )
     .acquire(checkEnv)
     .acquire(setupMongo)
+    .acquire(setupAuthorization)
     .acquire(setupInternals)
     .acquire(ensureIndexes)
     .acquire(setupListeners)
